Support for -GPT disks with Fat/EXFat volumes
WIP (More PIP) - Start support for GPT disks

Created a couple of new structures in the FS Structure header file, using same type of defines as the other structure.

Updated both fatlib and exfatlib to detect the GPT guard MBR entry and then walk through to header and then through the entries to see if the partition has the Microsoft Basic Data Partition and then tries to connect to that one.

So far appears to work.  I have not updated the Format code, in this project as today it totally blasts away all partition information.

Have WIP in our PFSLib of USBHost_t36 portotype that tries to do this on a partition.  Still testing.